+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 # Change log
 
+## v0.4.5
+
+### Main changes
+
+- harp_data can also be a new harp-python style folder.
+- Add NeuropixelData dataset for onix npx data
+- Add `BonsaiData` dataset for `.bonsai` workflow files.
+- Add `ignore_folders` option to `create_yaml_dict` to ignore specified folders
+  when parsing a directory. Hidden folders are always ignored.
+
+
+### Bugfix
+
+- Handle camera with _ in name correctly
+
+### Minor
+
+- GUI improvement. Clearer error message
+
 ## v0.4.4
 
 ### Bugfix